separated socket routes and much much more
