--- CircuitPython/HWconfig_template.json
@@ -1,56 +1,0 @@
-{
-    "Check-Hardware": ["Drivetrains", "IMU", "GPS"],
-    "Drivetrains":[
-        {
-            "motors":[
-                {
-                    "driver":"PhasedMotor",
-                    "address":"22,13"
-                },
-                {
-                    "driver":"PhasedMotor",
-                    "address":"17,18"
-                }
-            ],
-            "type":"Tank",
-            "max speed":"85"
-        },
-        {
-            "motors":[
-                {
-                    "driver":"NRF24L01",
-                    "address":"5,8",
-                    "name":"rfpi0"
-                }
-            ],
-            "type":"External"
-        }
-    ],
-    "IMU":[
-            {
-                "driver":"LSM9DS1",
-                "DoF":"9",
-                "address":"0x1c,0x6a"
-            },
-            {
-                "driver":"MPU6050",
-                "DoF":"6",
-                "address":"0x68"
-            },
-            {
-                "driver":"MAG3110",
-                "DoF":"3",
-                "address":"/dev/ttyUSB0"
-            }
-    ],
-    "GPS":[
-        {
-            "driver":"GPSserial",
-            "address":"/dev/ttyS0"
-        },
-        {
-            "driver":"GPSserial",
-            "address":"/dev/ttyACM0"
-        }
-    ]
-}
