+++ Python/bash_scripts/ask_pass_before_bash.sh
