+++ Python/common.py
@@ -1,4 +1,3 @@
 import board
-
 SPI_BUS = board.SPI()
 I2C_BUS = board.I2C()
